+++ tests/e03_human_genomes_test.py
@@ -18,8 +18,8 @@ if not skip:
         specific sequence.
         """
         tmp = mkdtemp()
-        genomepy.install_genome("hg38", "UCSC", genome_dir=tmp)
-        g = genomepy.Genome("hg38", genome_dir=tmp)
+        genomepy.install_genome("hg38", "UCSC", genomes_dir=tmp)
+        g = genomepy.Genome("hg38", genomes_dir=tmp)
         seq = g["chr6"][166168664:166168679]
         assert str(seq) == "CCTCCTCGCTCTCTT"
         shutil.rmtree(tmp)
@@ -32,8 +32,8 @@ if not skip:
         specific sequence.
         """
         tmp = mkdtemp()
-        genomepy.install_genome("GRCh38.p13", "Ensembl", genome_dir=tmp)
-        g = genomepy.Genome("GRCh38.p13", genome_dir=tmp)
+        genomepy.install_genome("GRCh38.p13", "Ensembl", genomes_dir=tmp)
+        g = genomepy.Genome("GRCh38.p13", genomes_dir=tmp)
         seq = g["6"][166168664:166168679]
         assert str(seq) == "CCTCCTCGCTCTCTT"
         shutil.rmtree(tmp)
@@ -46,8 +46,8 @@ if not skip:
         specific sequence.
         """
         tmp = mkdtemp()
-        genomepy.install_genome("GRCh38.p13", "NCBI", genome_dir=tmp)
-        g = genomepy.Genome("GRCh38.p13", genome_dir=tmp)
+        genomepy.install_genome("GRCh38.p13", "NCBI", genomes_dir=tmp)
+        g = genomepy.Genome("GRCh38.p13", genomes_dir=tmp)
         seq = g["6"][166168664:166168679]
         assert str(seq) == "CCTCCTCGCTCTCTT"
         shutil.rmtree(tmp)